Create proper UI components in conventional locations
- Add Text, Heading, Container, Stack, and Grid components to ui/ directory
- Follow Goose component architecture with proper exports and TypeScript types
- Update GooseComponentLibrary to import from proper component locations
- Remove inline component definitions for cleaner separation of concerns
