+++ ui/desktop/src/components/ui/stack.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+import { cn } from '../../utils';
+
+/**
+ * Stack - Flexbox stack layout
+ */
+interface StackProps {
+  children: React.ReactNode;
+  direction?: 'vertical' | 'horizontal';
+  gap?: 'none' | 'sm' | 'md' | 'lg';
+  align?: 'start' | 'center' | 'end' | 'stretch';
+  justify?: 'start' | 'center' | 'end' | 'between' | 'around';
+  className?: string;
+}
+
+const Stack: React.FC<StackProps> = ({ 
+  children, 
+  direction = 'vertical', 
+  gap = 'md',
+  align = 'stretch',
+  justify = 'start',
+  className 
+}) => {
+  const directionClasses = {
+    vertical: 'flex-col',
+    horizontal: 'flex-row'
+  };
+
+  const gapClasses = {
+    none: '',
+    sm: 'gap-2',
+    md: 'gap-4',
+    lg: 'gap-6'
+  };
+
+  const alignClasses = {
+    start: 'items-start',
+    center: 'items-center',
+    end: 'items-end',
+    stretch: 'items-stretch'
+  };
+
+  const justifyClasses = {
+    start: 'justify-start',
+    center: 'justify-center',
+    end: 'justify-end',
+    between: 'justify-between',
+    around: 'justify-around'
+  };
+
+  return (
+    <div className={cn(
+      'flex',
+      directionClasses[direction],
+      gapClasses[gap],
+      alignClasses[align],
+      justifyClasses[justify],
+      className
+    )}>
+      {children}
+    </div>
+  );
+};
+
+export { Stack };
+export type { StackProps };
